refactor(captcha): move email verification status to auth and drop dead code
`VerificationStatus` (and its `GET /api/auth/verification-status` route)
lived in the captcha package as a leftover of the verification rename;
email verification now belongs to auth, which already owns VerifyEmail
and the token flows. Move the service method, handler and route into
auth, keeping the API path and response shape unchanged, and delete the
unreferenced sentinel errors (`ErrEmailAlreadyVerified`,
`ErrEmailServiceDisabled`, `ErrGenerateToken`,
`ErrSendVerificationEmail`) plus the now-unused
`FindUserByID`/`FindUserByToken` repository methods. Fix the stale
"verification domain" doc comments throughout the captcha package.
