+++ backend/internal/auth/handler_test.go
@@ -1,4 +1,4 @@
-package captcha
+package auth
 
 import (
 	"encoding/json"
@@ -6,6 +6,9 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	"github.com/vexgo-org/vexgo/backend/internal/captcha"
+	"github.com/vexgo-org/vexgo/backend/internal/mailer"
+	"github.com/vexgo-org/vexgo/backend/internal/middleware"
 	"github.com/vexgo-org/vexgo/backend/internal/model"
 
 	"github.com/gin-gonic/gin"
@@ -14,7 +17,7 @@ import (
 func TestGetVerificationStatus_UserContextUint(t *testing.T) {
 	gin.SetMode(gin.TestMode)
 
-	_, db := newTestService(t)
+	_, _, db := newTestService(t)
 	u := model.User{
 		Username:      "alice",
 		Email:         "alice@example.com",
@@ -25,14 +28,17 @@ func TestGetVerificationStatus_UserContextUint(t *testing.T) {
 		t.Fatalf("failed to seed user: %v", err)
 	}
 
-	h := NewHandler(Deps{DB: db})
+	h := NewHandler(Deps{
+		DB:        db,
+		JWTSecret: testJWTSecret,
+		Files:     &fakeFiles{},
+		Mailer:    mailer.NewService(mailer.Deps{DB: db}),
+		Captcha:   captcha.NewService(captcha.Deps{DB: db}),
+	})
 
-	// The middleware stores the user id as uint in the "user" context map.
-	// The handler used to assert `.(float64)`, which always failed and fell
-	// through to a 500 — this test locks in the uint assertion.
 	r := gin.New()
 	r.GET("/verification-status", func(c *gin.Context) {
-		c.Set("user", map[string]any{"id": u.ID, "username": u.Username, "role": u.Role})
+		c.Set(middleware.CtxUserIDKey, u.ID)
 		h.GetVerificationStatus(c)
 	})
 
@@ -59,12 +65,18 @@ func TestGetVerificationStatus_UserContextUint(t *testing.T) {
 func TestGetVerificationStatus_UserNotFound(t *testing.T) {
 	gin.SetMode(gin.TestMode)
 
-	_, db := newTestService(t)
-	h := NewHandler(Deps{DB: db})
+	_, _, db := newTestService(t)
+	h := NewHandler(Deps{
+		DB:        db,
+		JWTSecret: testJWTSecret,
+		Files:     &fakeFiles{},
+		Mailer:    mailer.NewService(mailer.Deps{DB: db}),
+		Captcha:   captcha.NewService(captcha.Deps{DB: db}),
+	})
 
 	r := gin.New()
 	r.GET("/verification-status", func(c *gin.Context) {
-		c.Set("user", map[string]any{"id": uint(99999), "username": "ghost", "role": model.RoleGuest})
+		c.Set(middleware.CtxUserIDKey, uint(99999))
 		h.GetVerificationStatus(c)
 	})
